Moved PeopleRow into its own file
